+++ cmd/os_other.go
@@ -8,6 +8,6 @@ func hideConsole() {}
 
 func showErrorDialog(string) {}
 
-func spawnDetachedIfNeeded() bool { return false }
+func spawnDetachedIfNeeded(string) bool { return false }
 
 func isTrayChild() bool { return false }